Update organization subscription price tests

+++ app/costasiella/schema/organization_subscription_price.py
@@ -109,7 +109,7 @@ class OrganizationSubscriptionPriceQuery(graphene.ObjectType):
 class CreateOrganizationSubscriptionPrice(graphene.relay.ClientIDMutation):
     class Input:
         organization_subscription = graphene.ID(required=True)
-        price = graphene.Float(required=True, default_value=0)
+        price = graphene.Decimal(required=True, default_value=0)
         finance_tax_rate = graphene.ID(required=True)
         date_start = graphene.types.datetime.Date(required=True)
         date_end = graphene.types.datetime.Date(required=False, default_value=None)
@@ -140,7 +140,7 @@ class CreateOrganizationSubscriptionPrice(graphene.relay.ClientIDMutation):
 class UpdateOrganizationSubscriptionPrice(graphene.relay.ClientIDMutation):
     class Input:
         id = graphene.ID(required=True)
-        price = graphene.Float(required=True, default_value=0)
+        price = graphene.Decimal(required=True, default_value=0)
         finance_tax_rate = graphene.ID(required=True)
         date_start = graphene.types.datetime.Date(required=True)
         date_end = graphene.types.datetime.Date(required=False, default_value=None)
